requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
bleach==3.1.1
boto3==1.14.20
botocore==1.17.20
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.8
django-jsonfield==1.4.0
django-randomslugfield==0.3.0
django-storages==1.9.1
docutils==0.15.2
future==0.18.2
idna==2.10
importlib-metadata==1.6.1
jeepney==0.4.2
jmespath==0.10.0
keyring==21.1.0
kombu==4.6.11
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.5.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
readme-renderer==24.0
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
SecretStorage==3.1.2
six==1.15.0
sqlparse==0.3.1
tqdm==4.47.0
twine==3.1.1
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0
sendgrid==6.4.8
django-cors-headers
